Change of how to receive the data from the API Server.
Instead of Observable<any>().subscribe(data => {})

I changed it to this.stockList!: Observable<any> = this.http.get("url");
this.stockList!: Stock, was changed to this.stockList: Observable<any>
This way, we can use async when displaying the data to the client side.
When using async, we can reduce the errors coming from the framework.